ci: web and node might run in parallel and collide on pjson moves

+++ scripts/build.mjs
@@ -65,6 +65,8 @@ fs.copyFileSync('./package.json', `./package.json.BAK`);
   });
 })();
 
-// restore the pjs
-fs.copyFileSync('./package.json.BAK', './package.json');
-fs.unlinkSync('./package.json.BAK');
+if (fs.existsSync('./package.json.BAK')) {
+  // restore the pjs
+  fs.copyFileSync('./package.json.BAK', './package.json');
+  fs.unlinkSync('./package.json.BAK');
+}
